{{cookiecutter.microservice_name}}/app_pkg/requirements.txt
pydantic==1.8.2
fastapi==0.68.0
uvicorn==0.14.0
gunicorn==20.0.4
pandas==1.1.5
psycopg2==2.8.6
sqlalchemy==1.4.20
pydantic==1.8.2
uvicorn[standard]==0.14.0
python-dotenv==0.19.0
redis==3.5.3
